% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarlyIntervention.R
\docType{data}
\name{EarlyIntervention}
\alias{EarlyIntervention}
\title{A dataset containing a hypothetical infection rate pattern}
\format{
An object of class \code{numeric} of length 100.
}
\usage{
EarlyIntervention
}
\description{
This is the early intervention rate pattern used in the paper
}
\details{
We use this for our "rt" parameter

@source Xie (2020)
}
\keyword{datasets}
