% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Active_Infections.R
\name{Sim_Active_Infections}
\alias{Sim_Active_Infections}
\title{Simulate Active Infections}
\usage{
Sim_Active_Infections(
  dat1 = StandardData,
  dat2 = EarlyIntervention,
  dat3 = LowerRate,
  MU1 = 4,
  MU2 = 3.6,
  size1 = 1,
  size2 = 0.9,
  parallel = FALSE,
  max_iter = 5
)
}
\arguments{
\item{dat1}{Standard infection pattern (vector) for black barplot}

\item{dat2}{Comparison infection pattern (vector) for blue barplot}

\item{dat3}{Comparison infectin pattern (vector) for red barplot}

\item{MU1}{expected time  (integer) to getting infected for 1st column}

\item{MU2}{expected time (integer) to getting infected for 2nd column}

\item{size1}{Dispersion parameter (integer) for negative binomial distribution so that var = MU + MU^2/size (first column)}

\item{size2}{Dispersion parameter (integer) for negative binomial distribution so that var = MU + MU^2/size (second column)}

\item{parallel}{Option to parallelize simulation replicates}

\item{max_iter}{number if iterations for replicates}
}
\value{
2 x 3 barplots comparing daily active infections for different infection rates and different parameters (MU and size).
}
\description{
Simulate Active Infections
}
\examples{
Sim_Active_Infections(dat1 = StandardData,dat2 = EarlyIntervention,dat3 = LowerRate,MU1=4,MU2 = 3.6,size1=1,size2=0.9,parallel = FALSE,max_iter=5 )
}
