% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Daily_Hospitalizations.R
\name{Sim_Daily_Hospitalizations}
\alias{Sim_Daily_Hospitalizations}
\title{Simulate Daily Hospitalizations}
\usage{
Sim_Daily_Hospitalizations(
  MU1 = 4,
  MU2 = 3.6,
  dat1 = StandardData,
  dat2 = EarlyIntervention,
  dat3 = LowerRate,
  size1 = 1,
  size2 = 0.9,
  parallel = FALSE,
  max_iter = 5
)
}
\arguments{
\item{MU1}{expected time  (integer) to getting infected for 1st column}

\item{MU2}{expected time (integer) to getting infected for 2nd column}

\item{dat1}{Standard infection pattern (vector) for black barplot}

\item{dat2}{Comparison infection pattern (vector) for blue barplot}

\item{dat3}{Comparison infectin pattern (vector) for red barplot}

\item{size1}{Dispersion parameter (integer) for negative binomial distribution so that var = MU + MU^2/size (first column)}

\item{size2}{Dispersion parameter (integer) for negative binomial distribution so that var = MU + MU^2/size (second column)}

\item{parallel}{Option to use parallel processing for running multiple functions w different parameters. Default is False.}

\item{max_iter}{number if iterations for replicates}
}
\value{
2 x 3 barplots comparing daily hospitalizations for different infectino rates and different parameters (MU and size).
}
\description{
Simulate Daily Hospitalizations
}
\examples{
Sim_Daily_Hospitalizations = function(MU1=4,MU2 = 3.6,dat1 = StandardData,dat2 = EarlyIntervention,dat3 = LowerRate,size1=1,size2=0.9,parallel = FALSE,max_iter = 5)
}
