% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardData.R
\docType{data}
\name{StandardData}
\alias{StandardData}
\title{A dataset containing a hypothetical infection rate pattern}
\format{
An object of class \code{numeric} of length 100.
}
\usage{
StandardData
}
\description{
This is the standard rate pattern used in the paper
}
\details{
We use this for our "rt" parameter

@source Xie (2020)
}
\keyword{datasets}
